Refactor timezone setup
Move the timezone setup into its own file to avoid duplication.

Refs UserOfficeProject/stfc-user-office-project#276
